+++ bundles/org.enterprisedomain.classmaker/src/org/enterprisedomain/classmaker/core/ClassMakerPlugin.java
@@ -30,7 +30,6 @@ import org.eclipse.core.runtime.jobs.Job;
 import org.eclipse.core.runtime.jobs.ProgressProvider;
 import org.eclipse.emf.codegen.util.CodeGenUtil;
 import org.eclipse.emf.common.EMFPlugin;
-import org.eclipse.emf.ecore.plugin.EcorePlugin;
 import org.eclipse.jface.operation.IRunnableWithProgress;
 import org.enterprisedomain.classmaker.ClassMakerService;
 import org.enterprisedomain.classmaker.impl.ClassMakerServiceImpl;
@@ -174,7 +173,7 @@ public class ClassMakerPlugin extends Plugin {
 	public static void runWithProgress(IRunnableWithProgress runnable)
 			throws InvocationTargetException, InterruptedException {
 		if (runner == null)
-			return;
+			runner = new DefaultRunnerWithProgress();
 		runner.run(runnable);
 	}
