+++ bundles/org.enterprisedomain.classmaker/src/org/enterprisedomain/classmaker/core/DefaultRunnerWithProgress.java
@@ -1,0 +1,14 @@
+package org.enterprisedomain.classmaker.core;
+
+import java.lang.reflect.InvocationTargetException;
+
+import org.eclipse.jface.operation.IRunnableWithProgress;
+
+public class DefaultRunnerWithProgress implements IRunnerWithProgress {
+
+	@Override
+	public void run(IRunnableWithProgress runnable) throws InvocationTargetException, InterruptedException {
+		runnable.run(ClassMakerPlugin.getProgressMonitor());
+	}
+
+}
