+++ bundles/org.enterprisedomain.workbench/META-INF/MANIFEST.MF
@@ -42,7 +42,8 @@ Require-Bundle: javax.inject,
  org.eclipse.emf.ecp.application.e3;bundle-version="1.26.0",
  org.eclipse.pde.ds.annotations;bundle-version="1.2.200",
  org.apache.felix.scr;bundle-version="2.1.24",
- org.eclipse.equinox.event;bundle-version="1.6.0"
+ org.eclipse.equinox.event;bundle-version="1.6.0",
+ org.eclipse.ui.workbench
 Bundle-RequiredExecutionEnvironment: JavaSE-11
 Automatic-Module-Name: org.enterprisedomain.workbench
 Import-Package: javax.annotation;version="1.2.0"
